[#2] rpc/client: Remove additional wrapper
