+++ rpc/client/util.go
@@ -1,4 +1,4 @@
-package grpc
+package client
 
 import (
 	"fmt"
